# Api/api/docs/swagger.yaml
definitions:
...
  model.Error:
    properties:
      message:
        type: string
    type: object
...
  model.Like:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      user_id:
        type: string
    type: object
  model.Likes:
    properties:
      likes:
        items:
          $ref: '#/definitions/model.Like'
        type: array
    type: object
...
  model.PutLikeReq:
    properties:
      post_id:
        type: string
      user_id:
        type: string
    type: object
...
  model.StandardErrorModel:
    properties:
      error:
        $ref: '#/definitions/model.Error'
    type: object
...
info:
  contact: {}
paths:
...
  /v1/like:
    post:
      consumes:
      - application/json
      description: This API for putting like
      parameters:
      - description: PutLike
        in: body
        name: like
        required: true
        schema:
          $ref: '#/definitions/model.PutLikeReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Like'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: Putting Like
      tags:
      - like
  /v1/like/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting like by id
      parameters:
      - description: Like_Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Like'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: GetLikeInfo
      tags:
      - like
  /v1/like/getAllLikeuser/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting all likes user
      parameters:
      - description: User_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Like'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: GetAllPostLikesUser
      tags:
      - like
  /v1/like/getPostLike/{id}:
    get:
      consumes:
      - application/json
      description: This API for getting post's like
      parameters:
      - description: Post_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Likes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      summary: GetPostLike
      tags:
      - like
  /v1/like/takeLike/{id}:
    delete:
      consumes:
      - application/json
      description: This API for taking like
      parameters:
      - description: Like_id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Like'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.StandardErrorModel'
      security:
      - BearerAuth: []
      summary: TakeLike
      tags:
      - like
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"